+++ crates/biome_css_analyze/src/lint/nursery.rs
@@ -12,6 +12,7 @@ pub mod no_invalid_position_at_import_rule;
 pub mod no_unknown_function;
 pub mod no_unknown_selector_pseudo_element;
 pub mod no_unknown_unit;
+pub mod no_unmatchable_anb_selector;
 pub mod use_generic_font_names;
 
 declare_group! {
@@ -28,6 +29,7 @@ declare_group! {
             self :: no_unknown_function :: NoUnknownFunction ,
             self :: no_unknown_selector_pseudo_element :: NoUnknownSelectorPseudoElement ,
             self :: no_unknown_unit :: NoUnknownUnit ,
+            self :: no_unmatchable_anb_selector :: NoUnmatchableAnbSelector ,
             self :: use_generic_font_names :: UseGenericFontNames ,
         ]
      }