+++ crates/biome_css_analyze/src/lint/nursery/no_invalid_position_at_import_rule.rs
@@ -1,6 +1,6 @@
 use biome_analyze::{context::RuleContext, declare_rule, Ast, Rule, RuleDiagnostic, RuleSource};
 use biome_console::markup;
-use biome_css_syntax::{AnyCssRule, CssRuleList, CssImportAtRule};
+use biome_css_syntax::{AnyCssRule, CssImportAtRule, CssRuleList};
 use biome_rowan::AstNode;
 
 declare_rule! {
@@ -45,26 +45,28 @@ impl Rule for NoInvalidPositionAtImportRule {
         let mut is_invalid_position = false;
 
         for rule in node {
-
             let any_css_at_rule = match rule {
                 AnyCssRule::CssAtRule(item) => item.rule().ok(),
-                _ => None
+                _ => None,
             };
 
             if let Some(any_css_at_rule) = any_css_at_rule {
+                // Ignore @charset, @layer
+                if any_css_at_rule.as_css_charset_at_rule().is_some() {
+                    continue;
+                }
+                if any_css_at_rule.as_css_layer_at_rule().is_some() {
+                    continue;
+                }
 
-                    // Ignore @charset, @layer
-                    if any_css_at_rule.as_css_charset_at_rule().is_some() { continue;}
-                    if any_css_at_rule.as_css_layer_at_rule().is_some() { continue;}
-
-                    let import_rule = any_css_at_rule.as_css_import_at_rule().cloned();
-                    if let Some(import_rule) = import_rule {
-                        if is_invalid_position {
-                            return Some(import_rule);
-                        }
-                    } else {
-                        is_invalid_position = true;
+                let import_rule = any_css_at_rule.as_css_import_at_rule().cloned();
+                if let Some(import_rule) = import_rule {
+                    if is_invalid_position {
+                        return Some(import_rule);
                     }
+                } else {
+                    is_invalid_position = true;
+                }
             } else {
                 is_invalid_position = true;
             }