+++ crates/biome_js_analyze/src/lint/correctness/no_unused_labels.rs
@@ -6,8 +6,8 @@ use biome_analyze::{
 use biome_console::markup;
 use biome_diagnostics::Applicability;
 use biome_js_syntax::{
-    AnyJsStatement, JsBreakStatement, JsContinueStatement, JsLabeledStatement, JsLanguage,
-    TextRange, WalkEvent,
+    AnyJsStatement, JsBreakStatement, JsContinueStatement, JsFileSource, JsLabeledStatement,
+    JsLanguage, TextRange, WalkEvent,
 };
 
 use biome_rowan::{AstNode, BatchMutationExt, Language, SyntaxNode, SyntaxResult, TokenText};
@@ -20,6 +20,8 @@ declare_rule! {
     /// Disallow unused labels.
     ///
     /// Labels that are declared and never used are most likely an error due to incomplete refactoring.
+    ///
+    /// The rule ignores reactive Svelte statements in Svelte components.
     ///
     /// ## Examples
     ///
@@ -50,6 +52,12 @@ declare_rule! {
     ///     DEV: assert(n >= 0);
     ///     return n;
     /// }
+    /// ```
+    ///
+    /// ```svelte
+    /// <script>
+    /// $: { /* reactive block */ }
+    /// </script>
     /// ```
     pub NoUnusedLabels {
         version: "1.0.0",
@@ -155,7 +163,17 @@ impl Rule for NoUnusedLabels {
     type Signals = Option<Self::State>;
     type Options = ();
 
-    fn run(_: &RuleContext<Self>) -> Option<Self::State> {
+    fn run(ctx: &RuleContext<Self>) -> Option<Self::State> {
+        let label = ctx.query().label_token().ok()?;
+        let label = label.text_trimmed();
+        if label == "$"
+            && ctx
+                .source_type::<JsFileSource>()
+                .as_embedding_kind()
+                .is_svelte()
+        {
+            return None;
+        }
         Some(())
     }