+++ CHANGELOG.md
@@ -23,6 +23,8 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
   Contributed by @tasshi-me
 
+- `biome migrate eslint` now correctly resolve `@scope/eslint-config` ([#2705](https://github.com/biomejs/biome/issues/2705)). Contributed by @Conaclos
+
 ### Configuration
 
 ### Editors
@@ -32,6 +34,10 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 ### JavaScript APIs
 
 ### Linter
+
+#### New features
+
+- Add [nursery/useExplicitLengthCheck](https://biomejs.dev/linter/rules/use-explicit-length-check/).
 
 #### Bug fixes
 
@@ -49,6 +55,18 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
   This is a regression introduced by [#2394](https://github.com/biomejs/biome/issues/2394).
   This regression makes `noRedeclare` panics on every conditional types with `infer` bindings.
+
+  Contributed by @Conaclos
+
+- [noUnusedLabels](https://biomejs.dev/linter/rules/no-unused-labels/) and [noConfusingLabels](https://biomejs.dev/linter/rules/no-confusing-labels/) now ignore svelte reactive statements ([#2571](https://github.com/biomejs/biome/issues/2571)).
+
+  The rules now ignore reactive Svelte blocks in Svelte components.
+
+  ```svelte
+  <script>
+  $: { /* reactive block */ }
+  </script>
+  ```
 
   Contributed by @Conaclos